app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Renan Santana <renan_ota@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"
opencv-python = "^4.5.5"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
